+++ lib-remoteconfig/src/remoteconfigjson.cpp
@@ -61,7 +61,7 @@ uint32_t json_get_version(char *pOutBuffer, const uint32_t nOutBufferSize) {
 
 uint32_t json_get_uptime(char *pOutBuffer, const uint32_t nOutBufferSize) {
 	const auto nUptime = Hardware::Get()->GetUpTime();
-	const auto nLength = static_cast<uint32_t>(snprintf(pOutBuffer, nOutBufferSize, "{\"uptime\":%u}\n", nUptime));
+	const auto nLength = static_cast<uint32_t>(snprintf(pOutBuffer, nOutBufferSize, "{\"uptime\":%u}\n", static_cast<unsigned int>(nUptime)));
 	return nLength;
 }
